fc: fix building with intermittency
Also add a note about intermittent batched matmul

+++ common/fc.cpp
@@ -59,7 +59,7 @@ int16_t* const weights_tmp = op_buffer;
 #if INTERMITTENT
 static void gemm_recovery(Model* model, const ParameterInfo *input[], ParameterInfo* output, const Node* node, CurNodeFlags* node_flags, const NodeFlags* orig_node_flags,
                           // loop indices
-                          uint16_t* tile_channel_offset, uint16_t* tile_channel_idx, uint16_t* tile_a_row_offset, uint16_t* tile_b_col_offset, uint16_t* extended_tile_b_col_offset, uint32_t* part_idx,
+                          uint16_t* tile_channel_offset, uint16_t* tile_channel_idx, uint16_t* tile_a_row_offset, uint16_t* tile_b_col_offset, uint16_t* extended_tile_b_col_offset, uint16_t* part_idx,
                           // for state representation
                           int16_t* offset, uint16_t* next_output_turning_point, uint8_t* output_turning_point_idx, SlotInfo** output_slot_info) {
     const ParameterInfo *B = input[1];
@@ -84,6 +84,8 @@ static void gemm_recovery(Model* model, const ParameterInfo *input[], ParameterI
         }
         output_len *= output->dims[dim_idx];
     }
+
+    // TODO: get part_idx
 
     *tile_channel_idx = first_unfinished_value_offset / output_len;
     *tile_channel_offset = (*tile_channel_idx) * node_flags->gemm.tile_channel;
